requirements.txt
streamlit
pandas
numpy
requests
scipy               # For norm, logistic, etc.
scikit-learn        # For regression, pipelines, scaling
statsmodels         # For logistic regression, GLMs, hypothesis testing
google-generativeai # For Google Gemini API integration